+++ ovxlib/src/kernel/cpu/reduceany_internal_cpu.c
@@ -87,7 +87,7 @@ DEF_KERNEL_EXECUTOR(_compute)
     int32_t  innerSize   = 1;
     int32_t  inner       = 0;
     int32_t  outer       = 0;
-    int32_t  any_result  = 0.0f;
+    int32_t  any_result  = 0;
 
     for(i = 0; i < _INPUT_NUM; i ++)
     {
@@ -112,14 +112,14 @@ DEF_KERNEL_EXECUTOR(_compute)
     status = vsi_nn_kernel_scalar_read_int32((vsi_nn_kernel_scalar_t)param[SCALAR_INPUT_AXIS], &axis);
     CHECK_STATUS_FAIL_GOTO(status, final );
 
-    for (i = 0; i < axis; i++)
+    for (i = 0; i < (uint32_t)axis; i++)
     {
         innerSize *= out_attr[0]->shape->data[i];
     }
 
     axisSize = out_attr[0]->shape->data[axis];
 
-    for (i = axis + 1; i < (int32_t)out_attr[0]->shape->size; i++)
+    for (i = (uint32_t)axis + 1; i < out_attr[0]->shape->size; i++)
     {
         outerSize *= out_attr[0]->shape->data[i];
     }
@@ -129,7 +129,7 @@ DEF_KERNEL_EXECUTOR(_compute)
         for ( inner = 0; inner < innerSize; ++inner)
         {
             any_result = (!!(f32_in_buffer[0][outer * axisSize * innerSize + inner]));
-            for (i = 1; i < axisSize; ++i)
+            for (i = 1; i < (uint32_t)axisSize; ++i)
             {
                 int32_t value = (!!(f32_in_buffer[0][(outer * axisSize + i) * innerSize + inner]));
                 any_result = any_result || value;
